Recently, there have been vile online attacks looking to remove Richard M. Stallman from the FSF board of directors for expressing his personal opinions.
I will not stand idly when an icon of this community is attacked. As a
symbol of protest against Red-Hat for joining that witch hunt, no RPM
packages of mtools will be supplied with future releases.

For a more detailed explanation of this position, see Debian's proposal E
 https://www.debian.org/vote/2021/vote_002#texte

and the petition supporting his re-instatement:
 https://rms-support-letter.github.io/